ingestion/requirements.txt
feedparser==6.0.11
aiohttp==3.9.5
redis==5.0.4
python-dateutil==2.9.0
beautifulsoup4==4.12.3
lxml==5.2.2
